Handling absence of start and stop codons
When there is no start codon in a sequence and that ORFs starting at a start codon is requested then get_orfs() returns an empty set.
If there is no stop codon in the sequence, then get_orfs() returns the three open reading frames from the beginning to the end of the sequence.