Index comparison methods
Added two methods to get common index keys or a list of keys that are in one Table instance but not in the other one. Lists of keys are returned in Table objects with column names identical to the self instance.
Required to overwrite set_index in order to return a Table object instead of a DataFrame.