Table as method container
Table is not an extension of DataFrame anymore as it is not really needed and would actually make maintenance more complicated. It is only a class to hold class methods for utilities for easier table manipulation based on DataFrame. Most of the needed functionalities are already imploemented in DataFrame but may only require some wrapping in order to make them more practical for table manipulation in particular situations.