TableTransform class
This class will handle table transformations. Objects of this class hold the original copy of the DataFrame to transform and a working copy. Transformations are applied to the working copy but conditions are tested against the original DataFrame to avoid interference.  Transformation methods systematically return the self object for easy chaining of transformations.