Fixed tuple handling in Table.implode()
Tuple elements were wrongly changed into list during the implosion process. Numpy.unique was therefore replaced as its flattening of arrays was responsible for this issue. Also, if a tuple is the single element in a cell, then it is kept as is instead of being split into a list as was the fixed behoviour.